{"schema":"decodex/commit/1","summary":"split project boundary modules","authority":"manual"}
